+++ echoframework/src/main/kotlin/org/echo/mobile/framework/service/internal/subscription/ContractSubscriptionManagerImpl.kt
@@ -4,9 +4,11 @@ import com.google.common.reflect.TypeToken
 import com.google.gson.Gson
 import com.google.gson.JsonElement
 import org.echo.mobile.framework.core.logger.internal.LoggerCoreComponent
-import org.echo.mobile.framework.model.Log
+import org.echo.mobile.framework.model.contract.ContractLog
+import org.echo.mobile.framework.model.contract.Log
 import org.echo.mobile.framework.model.contract.output.ContractAddressOutputValueType
 import org.echo.mobile.framework.model.contract.output.ContractOutputDecoder
+import org.echo.mobile.framework.model.contract.processType
 import org.echo.mobile.framework.service.ContractSubscriptionManager
 import org.echo.mobile.framework.service.UpdateListener
 import org.echo.mobile.framework.support.toJsonObject
@@ -19,14 +21,14 @@ import java.util.concurrent.ConcurrentHashMap
  */
 class ContractSubscriptionManagerImpl : ContractSubscriptionManager {
 
-    private val listeners = ConcurrentHashMap<String, MutableList<UpdateListener<List<Log>>>>()
+    private val listeners = ConcurrentHashMap<String, MutableList<UpdateListener<List<ContractLog>>>>()
 
     private val gson = Gson()
-    private val type = object : TypeToken<List<Log>>() {}.type
+    private val type = object : TypeToken<List<ContractLog>>() {}.type
 
     private val contractDecoder = ContractOutputDecoder()
 
-    override fun registerListener(id: String, listener: UpdateListener<List<Log>>) {
+    override fun registerListener(id: String, listener: UpdateListener<List<ContractLog>>) {
         val accountListeners = listeners[id]
 
         if (accountListeners == null) {
@@ -41,36 +43,44 @@ class ContractSubscriptionManagerImpl : ContractSubscriptionManager {
 
     override fun registered(id: String): Boolean = listeners.containsKey(id)
 
-    override fun removeListeners(id: String): MutableList<UpdateListener<List<Log>>>? =
+    override fun removeListeners(id: String): MutableList<UpdateListener<List<ContractLog>>>? =
         listeners.remove(id)
 
     override fun clear() = listeners.clear()
 
-    override fun notify(contractId: String, logs: List<Log>) {
+    override fun notify(contractId: String, logs: List<ContractLog>) {
         listeners[contractId]?.forEach { listener ->
             listener.onUpdate(logs)
         }
     }
 
-    override fun processEvent(event: String): Map<String, List<Log>> {
+    override fun processEvent(event: String): Map<String, List<ContractLog>> {
         try {
             val dataParam = getDataParam(event)?.asJsonArray
 
             val logsJsonArray = dataParam?.firstOrNull()?.asJsonArray
-            val logs = gson.fromJson<List<Log>>(logsJsonArray, type)
 
-            val logsMap = mutableMapOf<String, MutableList<Log>>()
-            logs.forEach { log ->
+            val logsMap = mutableMapOf<String, MutableList<ContractLog>>()
+            logsJsonArray?.forEach { logArray ->
+                val contractArray = logArray.asJsonArray
+                val type = contractArray[0]
+                val contractJson = logArray.asJsonArray[1]
+
+                val log = Log(type.asInt, contractJson.toString())
+
+                val contractLog = log.processType()
 
                 val decodedValues = contractDecoder.decode(
-                    log.address.toByteArray(),
+                    contractLog?.address!!.toByteArray(),
                     listOf(ContractAddressOutputValueType())
                 )
 
                 val address = decodedValues.first().value.toString()
 
-                logsMap[address]?.add(log) ?: let {
-                    logsMap[address] = mutableListOf(log)
+                contractLog.address= address
+
+                logsMap[address]?.add(contractLog) ?: let {
+                    logsMap[address] = mutableListOf(contractLog)
                 }
             }